// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface TransformationInput {
    context: Uint8Array;
    response: http_request_result;
}
export interface Game {
    assists: bigint;
    date: string;
    rebounds: bigint;
    threes: bigint;
    points: bigint;
    opponent: string;
}
export interface TransformationOutput {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface http_header {
    value: string;
    name: string;
}
export interface http_request_result {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface backendInterface {
    fetchLast5GamesAndStore(playerName: string, games: Array<Game>): Promise<void>;
    getLast5Games(playerName: string): Promise<Array<Game> | null>;
    storeLast5Games(playerName: string, games: Array<Game>): Promise<void>;
    transform(input: TransformationInput): Promise<TransformationOutput>;
}